Initial commit; Basic architecture.
No features; just basic architecture for the app's development. The
app's name is "Work", though that is subject to change.

Work has the ability to print its version, a help message, allow the
user to select a custom configuration file, and the packaging logic is
implemented.

The selected license for this project is GPLv2.