Added the Task structure
The plan is to interact with Tasks in RST format; both for editing and
displaying. The Task class gives methods to allow for the creation of
Task instances via parsing RST text (which will be entered from a text
editor) and displaying the contents of the task in RST format.

These methods will parse the metadata for the task from the bottom of
the file as "key: value" pairs. Some of these metadata values are
treated specially. Any key that contains the strings "date" or "time"
will be parsed as dates, and the values for tags are expected to be
space or comma-separated and will be turned into a unique, sorted list.