Projects now perform their own task enforcement.
Rather than throw ugly error messages if a task was missing a critical
piece of metadata, the project will now corrct missing or bad attributes
when tasks are loaded into memory or about to be written to disk.

Tasks no longer have any mandated attributes; the project decides the
criterion for a task being "correct" via the _correctTask method.

This also removes a task's need to alphabetize its attributes; they'll
just be sorted when print.