feat: MASTER_11: Add comments to configuration in init
The config file edited by the user on project initialization will now
have useful comments shown. This was done by changing the initOnDisk
logic to copy a docstring to the .master.project file. This removes the
/etc/master.project.default file.

The default values for owner, project name, and prefix are now done by
replacing special templated values in the docstring. These initial ones
are __DEFAULT_OWNERS, __DEFAULT_PROJECT_NAME, and __DEFAULT_TASK_PREFIX

It is still possible to specify a custom config and if it has these
default values then they will be replaced automatically as well.