feat: MASTER_16: Add Time Support to Tasks
Added a few more special fields to Tasks when processing as ICS events.
start_time, due_time, and end_time. If these fields are present then the
task will include the time in the event's dtstart and dtend.

"duration" is a new field as well. Expected to be in the form of hours
and minutes. eg. 4 hours 30 minutes, or 4h 30m. Something like that.

The todo subcommand has a new output. Tasks are divided into 2
categories. If no time was specified then the command will list them as
to be performed sometime throughout the day, and tasks that have
specific times will be sorted by time. Their durations will also be
printed in days (if applicable), hours, and minutes.